Create FirstTemplate.Pde
This is just a sample template to see how to make images clickable and sort them properly. Also seen is the background(0,0,0) utilize as a change of page. The next time I touch this code , i may be utilizing objects/classes since the countries will share relatable information and it will make recoding simplier